correction to potential overflow fix

+++ src/lib/openjp2/pi.cpp
@@ -1268,7 +1268,7 @@ opj_pi_iterator_t *opj_pi_create_decode(opj_image_t *p_image,
 
     /* memory allocation for include */
 	l_current_pi->include = 00;
-	if (l_step_l && (l_tcp->numlayers < (UINT_MAX / l_step_l) - 1))		 {
+	if (l_step_l && (l_tcp->numlayers < (SIZE_MAX / l_step_l) - 1))		 {
 		l_current_pi->include = (int16_t*)opj_calloc((l_tcp->numlayers + 1) * l_step_l, sizeof(int16_t));
 	}
